blue-merle: restore old scrpit for SSH use
We restore from fa11fc45ca9191219c3965ad559ee8060f34a358.
We know that this worked well enough the last time as we run out of time
to test new things.
Ideally, though, we manage to consolidate the SSH use and the switch use
(and the Web use) to not have different implementations of the same
functionality so that testing gets much simpler.